Support config.extend() from es6 modules
As of babel 6, the default export from an es6 module is not assigned
directly to module.exports, instead to module.exports.default. As a
result, code require()ing es6 modules gets an object like
{default: foo}.

To fix this, a Config subclass has been created which uses a
ConfigFactory which knows about the default export of es6 modules.